Flink Delta Source PR_6 - Package Refactoring (#306)
