+++ flink-connector/src/main/java/io/delta/flink/source/internal/enumerator/monitor/TableMonitor.java
@@ -1,4 +1,4 @@
-package io.delta.flink.source.internal.enumerator;
+package io.delta.flink.source.internal.enumerator.monitor;
 
 import java.util.concurrent.Callable;
