+++ flink-connector/src/main/java/io/delta/flink/source/internal/enumerator/processor/SnapshotAndChangesTableProcessor.java
@@ -1,4 +1,4 @@
-package io.delta.flink.source.internal.enumerator;
+package io.delta.flink.source.internal.enumerator.processor;
 
 import java.util.Collection;
 import java.util.List;
